backend/requirements.txt
ai-edge-litert
librosa
numpy
matplotlib
fastapi
uvicorn[standard]
pyyaml
watchdog
soundfile
